WIP: Update MRI upstreams for Phosphorus
Adopt latest versions, namely;
- odlparent-9.0.2
- yangtools-7.0.2-SNAPSHOT
- infrautils-2.0.2
- mdsal-8.0.0-SNAPSHOT
- controller-4.0.0-SNAPSHOT
- aaa-0.14.0-SNAPSHOT
- netconf-2.0.0-SNAPSHOT

Also introduce model-{inventory,topology}, which were removed from
upstream.

+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/registry/flow/DeviceFlowRegistryImpl.java
@@ -62,7 +62,7 @@ public class DeviceFlowRegistryImpl implements DeviceFlowRegistry {
     private final Consumer<Flow> flowConsumer;
     private final DataBroker dataBroker;
 
-    public DeviceFlowRegistryImpl(final short version, final DataBroker dataBroker,
+    public DeviceFlowRegistryImpl(final Uint8 version, final DataBroker dataBroker,
             final KeyedInstanceIdentifier<Node, NodeKey> instanceIdentifier,
             final FlowGroupInfoHistoryAppender history) {
         this.dataBroker = requireNonNull(dataBroker);
